requirements.txt
aiogram
aiohttp
Pillow
python-dotenv
